+++ src/utils.js
@@ -1,0 +1,10 @@
+
+export const mapObjectToArray = (obj) => (
+    Object.entries(obj || {})
+        .map(([key, value]) => (
+            typeof value === 'object' ?
+                { ...value, key }
+                :
+                { key, value }
+        ))
+)
